requirements.txt
# Core Dependencies
google-generativeai==0.3.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# OCR Baseline (Optional - for comparison)
pytesseract==0.3.10
python-doctr==0.9.0

# Data Validation & Processing
pydantic==2.5.0
jsonschema==4.20.0
pandas==2.1.3

# Streamlit UI
streamlit==1.28.1
streamlit-image-comparison==0.0.3

# Evaluation & Metrics
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
